refactor: create utility function to make code DRY
refactor: created updateStudentDataInClass utility function to update student data in Firestore. This makes the codebase more 'DRY' and reduces the number of lines of code.
